validation/requirements.txt
jupyter
jupytext
matplotlib
pandas
psycopg2-binary
pytest
scipy
soundfile
sqlalchemy
